apps/memory_api/requirements-base.txt
# Base requirements for RAE Memory API (Lightweight)
# This file contains only essential dependencies without heavy ML libraries
# Total size: ~500MB vs 3-5GB with ML dependencies

# Core API Framework
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings

# Database & Vector Store
asyncpg
qdrant-client==1.10.1
pgvector
alembic

# HTTP & Networking
httpx==0.27.2

# Background Tasks
celery
redis

# Observability
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
prometheus-fastapi-instrumentator
prometheus_client
structlog

# LLM Providers (API-based, no local models)
litellm
openai
anthropic
google-generativeai

# Utilities
tenacity  # Retry logic
python-jose[cryptography]  # JWT handling
zstandard  # Compression
